+++ LaravelProject/database/migrations/2018_12_06_070926_create_allowance_table.php
@@ -17,10 +17,10 @@ class CreateAllowanceTable extends Migration
             $table->increments('id');
             $table->timestamps();
             $table->integer('emp_id');
-            $table->integer('overtime_a');
-            $table->integer('late_overtime_a');
+            $table->decimal('overtime_a');
+            $table->decimal('late_overtime_a');
             $table->integer('tardiness');
-            $table->integer('absence');
+            $table->decimal('absence');
         });
     }
